Refactoring the constants class to be a Input/Output messages class
